requirements.txt
boto3==1.28.44   
dj-database-url==2.0.0
Django==4.2.3
django-cors-headers==4.2.0
django-storages==1.14
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
mysqlclient==2.2.0
Pillow==10.0.0
python-decouple==3.8